requirements.txt
opencv-python>=4.8.0
numpy>=1.24.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=10.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
ftfy>=6.1.0
regex>=2023.0.0
tqdm>=4.65.0
git+https://github.com/openai/CLIP.git
